js file usage, destination recommendation
I have started to use js file, so i moved my js script from my htm to
the own js file and linked together and i built in a button which
reccomend a destination if you click it.
I am going to develop this function soon :) it will be more complex.